+++ User.h
@@ -13,7 +13,7 @@ struct Players
 
 bool verify(string email);
 bool comparePlayers(const Players &p1, const Players &p2);
-long int fileSize(string file);
+long long unsigned int fileSize(string file);
 void pushRecord(Players p);
 void printLeaderboard();
